NIFI-12265 Consolidated formatting in KeyIdentifierConverter
