+++ nifi-nar-bundles/nifi-pgp-bundle/nifi-pgp-service-api/src/test/java/org/apache/nifi/pgp/service/api/KeyIdentifierConverterTest.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.nifi.processors.pgp.io;
+package org.apache.nifi.pgp.service.api;
 
 import org.junit.jupiter.api.Test;
